Predefined device definitions

AtmelToolsDeviceFiles comes with some pre-defined devices files. They are stored as lazy Artifacts downloaded from https://packs.download.microchip.com that can be summoned at will.

For convenience, they are accessible through the AtmelToolsDeviceFiles.AtmelLibrary sub-module.

Adding new devices.

More device files are available in Microship's download page and can be added uppon request. Open an issue if you require one!

Accessible device files

AtmelToolsDeviceFiles.AtmelLibrary.AT90CAN128

AtmelToolsDeviceFiles.AtmelLibrary.AT90CAN32

AtmelToolsDeviceFiles.AtmelLibrary.AT90CAN64

AtmelToolsDeviceFiles.AtmelLibrary.AT90PWM1

AtmelToolsDeviceFiles.AtmelLibrary.AT90PWM161

AtmelToolsDeviceFiles.AtmelLibrary.AT90PWM216

AtmelToolsDeviceFiles.AtmelLibrary.AT90PWM2B

AtmelToolsDeviceFiles.AtmelLibrary.AT90PWM3

AtmelToolsDeviceFiles.AtmelLibrary.AT90PWM316

AtmelToolsDeviceFiles.AtmelLibrary.AT90PWM3B

AtmelToolsDeviceFiles.AtmelLibrary.AT90PWM81

AtmelToolsDeviceFiles.AtmelLibrary.AT90USB1286

AtmelToolsDeviceFiles.AtmelLibrary.AT90USB1287

AtmelToolsDeviceFiles.AtmelLibrary.AT90USB162

AtmelToolsDeviceFiles.AtmelLibrary.AT90USB646

AtmelToolsDeviceFiles.AtmelLibrary.AT90USB647

AtmelToolsDeviceFiles.AtmelLibrary.AT90USB82

AtmelToolsDeviceFiles.AtmelLibrary.ATmega128

AtmelToolsDeviceFiles.AtmelLibrary.ATmega1280

AtmelToolsDeviceFiles.AtmelLibrary.ATmega1281

AtmelToolsDeviceFiles.AtmelLibrary.ATmega1284

AtmelToolsDeviceFiles.AtmelLibrary.ATmega1284P

AtmelToolsDeviceFiles.AtmelLibrary.ATmega1284RFR2

AtmelToolsDeviceFiles.AtmelLibrary.ATmega128A

AtmelToolsDeviceFiles.AtmelLibrary.ATmega128RFA1

AtmelToolsDeviceFiles.AtmelLibrary.ATmega128RFR2

AtmelToolsDeviceFiles.AtmelLibrary.ATmega16

AtmelToolsDeviceFiles.AtmelLibrary.ATmega1608

AtmelToolsDeviceFiles.AtmelLibrary.ATmega1609

AtmelToolsDeviceFiles.AtmelLibrary.ATmega162

AtmelToolsDeviceFiles.AtmelLibrary.ATmega164A

AtmelToolsDeviceFiles.AtmelLibrary.ATmega164P

AtmelToolsDeviceFiles.AtmelLibrary.ATmega164PA

AtmelToolsDeviceFiles.AtmelLibrary.ATmega165A

AtmelToolsDeviceFiles.AtmelLibrary.ATmega165P

AtmelToolsDeviceFiles.AtmelLibrary.ATmega165PA

AtmelToolsDeviceFiles.AtmelLibrary.ATmega168

AtmelToolsDeviceFiles.AtmelLibrary.ATmega168A

AtmelToolsDeviceFiles.AtmelLibrary.ATmega168P

AtmelToolsDeviceFiles.AtmelLibrary.ATmega168PA

AtmelToolsDeviceFiles.AtmelLibrary.ATmega168PB

AtmelToolsDeviceFiles.AtmelLibrary.ATmega169A

AtmelToolsDeviceFiles.AtmelLibrary.ATmega169P

AtmelToolsDeviceFiles.AtmelLibrary.ATmega169PA

AtmelToolsDeviceFiles.AtmelLibrary.ATmega16A

AtmelToolsDeviceFiles.AtmelLibrary.ATmega16HVA

AtmelToolsDeviceFiles.AtmelLibrary.ATmega16HVB

AtmelToolsDeviceFiles.AtmelLibrary.ATmega16HVBrevB

AtmelToolsDeviceFiles.AtmelLibrary.ATmega16M1

AtmelToolsDeviceFiles.AtmelLibrary.ATmega16U2

AtmelToolsDeviceFiles.AtmelLibrary.ATmega16U4

AtmelToolsDeviceFiles.AtmelLibrary.ATmega2560

AtmelToolsDeviceFiles.AtmelLibrary.ATmega2561

AtmelToolsDeviceFiles.AtmelLibrary.ATmega2564RFR2

AtmelToolsDeviceFiles.AtmelLibrary.ATmega256RFR2

AtmelToolsDeviceFiles.AtmelLibrary.ATmega32

AtmelToolsDeviceFiles.AtmelLibrary.ATmega3208

AtmelToolsDeviceFiles.AtmelLibrary.ATmega3209

AtmelToolsDeviceFiles.AtmelLibrary.ATmega324A

AtmelToolsDeviceFiles.AtmelLibrary.ATmega324P

AtmelToolsDeviceFiles.AtmelLibrary.ATmega324PA

AtmelToolsDeviceFiles.AtmelLibrary.ATmega324PB

AtmelToolsDeviceFiles.AtmelLibrary.ATmega325

AtmelToolsDeviceFiles.AtmelLibrary.ATmega3250

AtmelToolsDeviceFiles.AtmelLibrary.ATmega3250A

AtmelToolsDeviceFiles.AtmelLibrary.ATmega3250P

AtmelToolsDeviceFiles.AtmelLibrary.ATmega3250PA

AtmelToolsDeviceFiles.AtmelLibrary.ATmega325A

AtmelToolsDeviceFiles.AtmelLibrary.ATmega325P

AtmelToolsDeviceFiles.AtmelLibrary.ATmega325PA

AtmelToolsDeviceFiles.AtmelLibrary.ATmega328

AtmelToolsDeviceFiles.AtmelLibrary.ATmega328P

AtmelToolsDeviceFiles.AtmelLibrary.ATmega328PB

AtmelToolsDeviceFiles.AtmelLibrary.ATmega329

AtmelToolsDeviceFiles.AtmelLibrary.ATmega3290

AtmelToolsDeviceFiles.AtmelLibrary.ATmega3290A

AtmelToolsDeviceFiles.AtmelLibrary.ATmega3290P

AtmelToolsDeviceFiles.AtmelLibrary.ATmega3290PA

AtmelToolsDeviceFiles.AtmelLibrary.ATmega329A

AtmelToolsDeviceFiles.AtmelLibrary.ATmega329P

AtmelToolsDeviceFiles.AtmelLibrary.ATmega329PA

AtmelToolsDeviceFiles.AtmelLibrary.ATmega32A

AtmelToolsDeviceFiles.AtmelLibrary.ATmega32C1

AtmelToolsDeviceFiles.AtmelLibrary.ATmega32HVB

AtmelToolsDeviceFiles.AtmelLibrary.ATmega32HVBrevB

AtmelToolsDeviceFiles.AtmelLibrary.ATmega32M1

AtmelToolsDeviceFiles.AtmelLibrary.ATmega32U2

AtmelToolsDeviceFiles.AtmelLibrary.ATmega32U4

AtmelToolsDeviceFiles.AtmelLibrary.ATmega406

AtmelToolsDeviceFiles.AtmelLibrary.ATmega48

AtmelToolsDeviceFiles.AtmelLibrary.ATmega4808

AtmelToolsDeviceFiles.AtmelLibrary.ATmega4809

AtmelToolsDeviceFiles.AtmelLibrary.ATmega48A

AtmelToolsDeviceFiles.AtmelLibrary.ATmega48P

AtmelToolsDeviceFiles.AtmelLibrary.ATmega48PA

AtmelToolsDeviceFiles.AtmelLibrary.ATmega48PB

AtmelToolsDeviceFiles.AtmelLibrary.ATmega64

AtmelToolsDeviceFiles.AtmelLibrary.ATmega640

AtmelToolsDeviceFiles.AtmelLibrary.ATmega644

AtmelToolsDeviceFiles.AtmelLibrary.ATmega644A

AtmelToolsDeviceFiles.AtmelLibrary.ATmega644P

AtmelToolsDeviceFiles.AtmelLibrary.ATmega644PA

AtmelToolsDeviceFiles.AtmelLibrary.ATmega644RFR2

AtmelToolsDeviceFiles.AtmelLibrary.ATmega645

AtmelToolsDeviceFiles.AtmelLibrary.ATmega6450

AtmelToolsDeviceFiles.AtmelLibrary.ATmega6450A

AtmelToolsDeviceFiles.AtmelLibrary.ATmega6450P

AtmelToolsDeviceFiles.AtmelLibrary.ATmega645A

AtmelToolsDeviceFiles.AtmelLibrary.ATmega645P

AtmelToolsDeviceFiles.AtmelLibrary.ATmega649

AtmelToolsDeviceFiles.AtmelLibrary.ATmega6490

AtmelToolsDeviceFiles.AtmelLibrary.ATmega6490A

AtmelToolsDeviceFiles.AtmelLibrary.ATmega6490P

AtmelToolsDeviceFiles.AtmelLibrary.ATmega649A

AtmelToolsDeviceFiles.AtmelLibrary.ATmega649P

AtmelToolsDeviceFiles.AtmelLibrary.ATmega64A

AtmelToolsDeviceFiles.AtmelLibrary.ATmega64C1

AtmelToolsDeviceFiles.AtmelLibrary.ATmega64HVE2

AtmelToolsDeviceFiles.AtmelLibrary.ATmega64M1

AtmelToolsDeviceFiles.AtmelLibrary.ATmega64RFR2

AtmelToolsDeviceFiles.AtmelLibrary.ATmega8

AtmelToolsDeviceFiles.AtmelLibrary.ATmega808

AtmelToolsDeviceFiles.AtmelLibrary.ATmega809

AtmelToolsDeviceFiles.AtmelLibrary.ATmega8515

AtmelToolsDeviceFiles.AtmelLibrary.ATmega8535

AtmelToolsDeviceFiles.AtmelLibrary.ATmega88

AtmelToolsDeviceFiles.AtmelLibrary.ATmega88A

AtmelToolsDeviceFiles.AtmelLibrary.ATmega88P

AtmelToolsDeviceFiles.AtmelLibrary.ATmega88PA

AtmelToolsDeviceFiles.AtmelLibrary.ATmega88PB

AtmelToolsDeviceFiles.AtmelLibrary.ATmega8A

AtmelToolsDeviceFiles.AtmelLibrary.ATmega8HVA

AtmelToolsDeviceFiles.AtmelLibrary.ATmega8U2

AtmelToolsDeviceFiles.AtmelLibrary.ATtiny10

AtmelToolsDeviceFiles.AtmelLibrary.ATtiny102

AtmelToolsDeviceFiles.AtmelLibrary.ATtiny104

AtmelToolsDeviceFiles.AtmelLibrary.ATtiny11

AtmelToolsDeviceFiles.AtmelLibrary.ATtiny12

AtmelToolsDeviceFiles.AtmelLibrary.ATtiny13

AtmelToolsDeviceFiles.AtmelLibrary.ATtiny13A

AtmelToolsDeviceFiles.AtmelLibrary.ATtiny15

AtmelToolsDeviceFiles.AtmelLibrary.ATtiny1604

AtmelToolsDeviceFiles.AtmelLibrary.ATtiny1606

AtmelToolsDeviceFiles.AtmelLibrary.ATtiny1607

AtmelToolsDeviceFiles.AtmelLibrary.ATtiny1614

AtmelToolsDeviceFiles.AtmelLibrary.ATtiny1616

AtmelToolsDeviceFiles.AtmelLibrary.ATtiny1617

AtmelToolsDeviceFiles.AtmelLibrary.ATtiny1624

AtmelToolsDeviceFiles.AtmelLibrary.ATtiny1626

AtmelToolsDeviceFiles.AtmelLibrary.ATtiny1627

AtmelToolsDeviceFiles.AtmelLibrary.ATtiny1634

AtmelToolsDeviceFiles.AtmelLibrary.ATtiny167

AtmelToolsDeviceFiles.AtmelLibrary.ATtiny20

AtmelToolsDeviceFiles.AtmelLibrary.ATtiny202

AtmelToolsDeviceFiles.AtmelLibrary.ATtiny204

AtmelToolsDeviceFiles.AtmelLibrary.ATtiny212

AtmelToolsDeviceFiles.AtmelLibrary.ATtiny214

AtmelToolsDeviceFiles.AtmelLibrary.ATtiny2313

AtmelToolsDeviceFiles.AtmelLibrary.ATtiny2313A

AtmelToolsDeviceFiles.AtmelLibrary.ATtiny24

AtmelToolsDeviceFiles.AtmelLibrary.ATtiny24A

AtmelToolsDeviceFiles.AtmelLibrary.ATtiny25

AtmelToolsDeviceFiles.AtmelLibrary.ATtiny26

AtmelToolsDeviceFiles.AtmelLibrary.ATtiny261

AtmelToolsDeviceFiles.AtmelLibrary.ATtiny261A

AtmelToolsDeviceFiles.AtmelLibrary.ATtiny3216

AtmelToolsDeviceFiles.AtmelLibrary.ATtiny3217

AtmelToolsDeviceFiles.AtmelLibrary.ATtiny3224

AtmelToolsDeviceFiles.AtmelLibrary.ATtiny3226

AtmelToolsDeviceFiles.AtmelLibrary.ATtiny3227

AtmelToolsDeviceFiles.AtmelLibrary.ATtiny4

AtmelToolsDeviceFiles.AtmelLibrary.ATtiny40

AtmelToolsDeviceFiles.AtmelLibrary.ATtiny402

AtmelToolsDeviceFiles.AtmelLibrary.ATtiny404

AtmelToolsDeviceFiles.AtmelLibrary.ATtiny406

AtmelToolsDeviceFiles.AtmelLibrary.ATtiny412

AtmelToolsDeviceFiles.AtmelLibrary.ATtiny414

AtmelToolsDeviceFiles.AtmelLibrary.ATtiny416

AtmelToolsDeviceFiles.AtmelLibrary.ATtiny417

AtmelToolsDeviceFiles.AtmelLibrary.ATtiny424

AtmelToolsDeviceFiles.AtmelLibrary.ATtiny426

AtmelToolsDeviceFiles.AtmelLibrary.ATtiny427

AtmelToolsDeviceFiles.AtmelLibrary.ATtiny4313

AtmelToolsDeviceFiles.AtmelLibrary.ATtiny43U

AtmelToolsDeviceFiles.AtmelLibrary.ATtiny44

AtmelToolsDeviceFiles.AtmelLibrary.ATtiny441

AtmelToolsDeviceFiles.AtmelLibrary.ATtiny44A

AtmelToolsDeviceFiles.AtmelLibrary.ATtiny45

AtmelToolsDeviceFiles.AtmelLibrary.ATtiny461

AtmelToolsDeviceFiles.AtmelLibrary.ATtiny461A

AtmelToolsDeviceFiles.AtmelLibrary.ATtiny48

AtmelToolsDeviceFiles.AtmelLibrary.ATtiny5

AtmelToolsDeviceFiles.AtmelLibrary.ATtiny804

AtmelToolsDeviceFiles.AtmelLibrary.ATtiny806

AtmelToolsDeviceFiles.AtmelLibrary.ATtiny807

AtmelToolsDeviceFiles.AtmelLibrary.ATtiny814

AtmelToolsDeviceFiles.AtmelLibrary.ATtiny816

AtmelToolsDeviceFiles.AtmelLibrary.ATtiny817

AtmelToolsDeviceFiles.AtmelLibrary.ATtiny824

AtmelToolsDeviceFiles.AtmelLibrary.ATtiny826

AtmelToolsDeviceFiles.AtmelLibrary.ATtiny827

AtmelToolsDeviceFiles.AtmelLibrary.ATtiny828

AtmelToolsDeviceFiles.AtmelLibrary.ATtiny84

AtmelToolsDeviceFiles.AtmelLibrary.ATtiny841

AtmelToolsDeviceFiles.AtmelLibrary.ATtiny84A

AtmelToolsDeviceFiles.AtmelLibrary.ATtiny85

AtmelToolsDeviceFiles.AtmelLibrary.ATtiny861

AtmelToolsDeviceFiles.AtmelLibrary.ATtiny861A

AtmelToolsDeviceFiles.AtmelLibrary.ATtiny87

AtmelToolsDeviceFiles.AtmelLibrary.ATtiny88

AtmelToolsDeviceFiles.AtmelLibrary.ATtiny9

AtmelToolsDeviceFiles.AtmelLibrary — Module

Contains paths to Microship's provided device definition files (lazily downloaded as artifacts).

AtmelToolsDeviceFiles.AtmelLibrary.AT90CAN128 — Method

The path of the AT90CAN128 file definition.

AtmelToolsDeviceFiles.AtmelLibrary.AT90CAN32 — Method

The path of the AT90CAN32 file definition.

AtmelToolsDeviceFiles.AtmelLibrary.AT90CAN64 — Method

The path of the AT90CAN64 file definition.

AtmelToolsDeviceFiles.AtmelLibrary.AT90PWM1 — Method

The path of the AT90PWM1 file definition.

AtmelToolsDeviceFiles.AtmelLibrary.AT90PWM161 — Method

The path of the AT90PWM161 file definition.

AtmelToolsDeviceFiles.AtmelLibrary.AT90PWM216 — Method

The path of the AT90PWM216 file definition.

AtmelToolsDeviceFiles.AtmelLibrary.AT90PWM2B — Method

The path of the AT90PWM2B file definition.

AtmelToolsDeviceFiles.AtmelLibrary.AT90PWM3 — Method

The path of the AT90PWM3 file definition.

AtmelToolsDeviceFiles.AtmelLibrary.AT90PWM316 — Method

The path of the AT90PWM316 file definition.

AtmelToolsDeviceFiles.AtmelLibrary.AT90PWM3B — Method

The path of the AT90PWM3B file definition.

AtmelToolsDeviceFiles.AtmelLibrary.AT90PWM81 — Method

The path of the AT90PWM81 file definition.

AtmelToolsDeviceFiles.AtmelLibrary.AT90USB1286 — Method

The path of the AT90USB1286 file definition.

AtmelToolsDeviceFiles.AtmelLibrary.AT90USB1287 — Method

The path of the AT90USB1287 file definition.

AtmelToolsDeviceFiles.AtmelLibrary.AT90USB162 — Method

The path of the AT90USB162 file definition.

AtmelToolsDeviceFiles.AtmelLibrary.AT90USB646 — Method

The path of the AT90USB646 file definition.

AtmelToolsDeviceFiles.AtmelLibrary.AT90USB647 — Method

The path of the AT90USB647 file definition.

AtmelToolsDeviceFiles.AtmelLibrary.AT90USB82 — Method

The path of the AT90USB82 file definition.

AtmelToolsDeviceFiles.AtmelLibrary.ATmega128 — Method

The path of the ATmega128 file definition.

AtmelToolsDeviceFiles.AtmelLibrary.ATmega1280 — Method

The path of the ATmega1280 file definition.

AtmelToolsDeviceFiles.AtmelLibrary.ATmega1281 — Method

The path of the ATmega1281 file definition.

AtmelToolsDeviceFiles.AtmelLibrary.ATmega1284 — Method

The path of the ATmega1284 file definition.

AtmelToolsDeviceFiles.AtmelLibrary.ATmega1284P — Method

The path of the ATmega1284P file definition.

AtmelToolsDeviceFiles.AtmelLibrary.ATmega1284RFR2 — Method

The path of the ATmega1284RFR2 file definition.

AtmelToolsDeviceFiles.AtmelLibrary.ATmega128A — Method

The path of the ATmega128A file definition.

AtmelToolsDeviceFiles.AtmelLibrary.ATmega128RFA1 — Method

The path of the ATmega128RFA1 file definition.

AtmelToolsDeviceFiles.AtmelLibrary.ATmega128RFR2 — Method

The path of the ATmega128RFR2 file definition.

AtmelToolsDeviceFiles.AtmelLibrary.ATmega16 — Method

The path of the ATmega16 file definition.

AtmelToolsDeviceFiles.AtmelLibrary.ATmega1608 — Method

The path of the ATmega1608 file definition.

AtmelToolsDeviceFiles.AtmelLibrary.ATmega1609 — Method

The path of the ATmega1609 file definition.

AtmelToolsDeviceFiles.AtmelLibrary.ATmega162 — Method

The path of the ATmega162 file definition.

AtmelToolsDeviceFiles.AtmelLibrary.ATmega164A — Method

The path of the ATmega164A file definition.

AtmelToolsDeviceFiles.AtmelLibrary.ATmega164P — Method

The path of the ATmega164P file definition.

AtmelToolsDeviceFiles.AtmelLibrary.ATmega164PA — Method

The path of the ATmega164PA file definition.

AtmelToolsDeviceFiles.AtmelLibrary.ATmega165A — Method

The path of the ATmega165A file definition.

AtmelToolsDeviceFiles.AtmelLibrary.ATmega165P — Method

The path of the ATmega165P file definition.

AtmelToolsDeviceFiles.AtmelLibrary.ATmega165PA — Method

The path of the ATmega165PA file definition.

AtmelToolsDeviceFiles.AtmelLibrary.ATmega168 — Method

The path of the ATmega168 file definition.

AtmelToolsDeviceFiles.AtmelLibrary.ATmega168A — Method

The path of the ATmega168A file definition.

AtmelToolsDeviceFiles.AtmelLibrary.ATmega168P — Method

The path of the ATmega168P file definition.

AtmelToolsDeviceFiles.AtmelLibrary.ATmega168PA — Method

The path of the ATmega168PA file definition.

AtmelToolsDeviceFiles.AtmelLibrary.ATmega168PB — Method

The path of the ATmega168PB file definition.

AtmelToolsDeviceFiles.AtmelLibrary.ATmega169A — Method

The path of the ATmega169A file definition.

AtmelToolsDeviceFiles.AtmelLibrary.ATmega169P — Method

The path of the ATmega169P file definition.

AtmelToolsDeviceFiles.AtmelLibrary.ATmega169PA — Method

The path of the ATmega169PA file definition.

AtmelToolsDeviceFiles.AtmelLibrary.ATmega16A — Method

The path of the ATmega16A file definition.

AtmelToolsDeviceFiles.AtmelLibrary.ATmega16HVA — Method

The path of the ATmega16HVA file definition.

AtmelToolsDeviceFiles.AtmelLibrary.ATmega16HVB — Method

The path of the ATmega16HVB file definition.

AtmelToolsDeviceFiles.AtmelLibrary.ATmega16HVBrevB — Method

The path of the ATmega16HVBrevB file definition.

AtmelToolsDeviceFiles.AtmelLibrary.ATmega16M1 — Method

The path of the ATmega16M1 file definition.

AtmelToolsDeviceFiles.AtmelLibrary.ATmega16U2 — Method

The path of the ATmega16U2 file definition.

AtmelToolsDeviceFiles.AtmelLibrary.ATmega16U4 — Method

The path of the ATmega16U4 file definition.

AtmelToolsDeviceFiles.AtmelLibrary.ATmega2560 — Method

The path of the ATmega2560 file definition.

AtmelToolsDeviceFiles.AtmelLibrary.ATmega2561 — Method

The path of the ATmega2561 file definition.

AtmelToolsDeviceFiles.AtmelLibrary.ATmega2564RFR2 — Method

The path of the ATmega2564RFR2 file definition.

AtmelToolsDeviceFiles.AtmelLibrary.ATmega256RFR2 — Method

The path of the ATmega256RFR2 file definition.

AtmelToolsDeviceFiles.AtmelLibrary.ATmega32 — Method

The path of the ATmega32 file definition.

AtmelToolsDeviceFiles.AtmelLibrary.ATmega3208 — Method

The path of the ATmega3208 file definition.

AtmelToolsDeviceFiles.AtmelLibrary.ATmega3209 — Method

The path of the ATmega3209 file definition.

AtmelToolsDeviceFiles.AtmelLibrary.ATmega324A — Method

The path of the ATmega324A file definition.

AtmelToolsDeviceFiles.AtmelLibrary.ATmega324P — Method

The path of the ATmega324P file definition.

AtmelToolsDeviceFiles.AtmelLibrary.ATmega324PA — Method

The path of the ATmega324PA file definition.

AtmelToolsDeviceFiles.AtmelLibrary.ATmega324PB — Method

The path of the ATmega324PB file definition.

AtmelToolsDeviceFiles.AtmelLibrary.ATmega325 — Method

The path of the ATmega325 file definition.

AtmelToolsDeviceFiles.AtmelLibrary.ATmega3250 — Method

The path of the ATmega3250 file definition.

AtmelToolsDeviceFiles.AtmelLibrary.ATmega3250A — Method

The path of the ATmega3250A file definition.

AtmelToolsDeviceFiles.AtmelLibrary.ATmega3250P — Method

The path of the ATmega3250P file definition.

AtmelToolsDeviceFiles.AtmelLibrary.ATmega3250PA — Method

The path of the ATmega3250PA file definition.

AtmelToolsDeviceFiles.AtmelLibrary.ATmega325A — Method

The path of the ATmega325A file definition.

AtmelToolsDeviceFiles.AtmelLibrary.ATmega325P — Method

The path of the ATmega325P file definition.

AtmelToolsDeviceFiles.AtmelLibrary.ATmega325PA — Method

The path of the ATmega325PA file definition.

AtmelToolsDeviceFiles.AtmelLibrary.ATmega328 — Method

The path of the ATmega328 file definition.

AtmelToolsDeviceFiles.AtmelLibrary.ATmega328P — Method

The path of the ATmega328P file definition.

AtmelToolsDeviceFiles.AtmelLibrary.ATmega328PB — Method

The path of the ATmega328PB file definition.

AtmelToolsDeviceFiles.AtmelLibrary.ATmega329 — Method

The path of the ATmega329 file definition.

AtmelToolsDeviceFiles.AtmelLibrary.ATmega3290 — Method

The path of the ATmega3290 file definition.

AtmelToolsDeviceFiles.AtmelLibrary.ATmega3290A — Method

The path of the ATmega3290A file definition.

AtmelToolsDeviceFiles.AtmelLibrary.ATmega3290P — Method

The path of the ATmega3290P file definition.

AtmelToolsDeviceFiles.AtmelLibrary.ATmega3290PA — Method

The path of the ATmega3290PA file definition.

AtmelToolsDeviceFiles.AtmelLibrary.ATmega329A — Method

The path of the ATmega329A file definition.

AtmelToolsDeviceFiles.AtmelLibrary.ATmega329P — Method

The path of the ATmega329P file definition.

AtmelToolsDeviceFiles.AtmelLibrary.ATmega329PA — Method

The path of the ATmega329PA file definition.

AtmelToolsDeviceFiles.AtmelLibrary.ATmega32A — Method

The path of the ATmega32A file definition.

AtmelToolsDeviceFiles.AtmelLibrary.ATmega32C1 — Method

The path of the ATmega32C1 file definition.

AtmelToolsDeviceFiles.AtmelLibrary.ATmega32HVB — Method

The path of the ATmega32HVB file definition.

AtmelToolsDeviceFiles.AtmelLibrary.ATmega32HVBrevB — Method

The path of the ATmega32HVBrevB file definition.

AtmelToolsDeviceFiles.AtmelLibrary.ATmega32M1 — Method

The path of the ATmega32M1 file definition.

AtmelToolsDeviceFiles.AtmelLibrary.ATmega32U2 — Method

The path of the ATmega32U2 file definition.

AtmelToolsDeviceFiles.AtmelLibrary.ATmega32U4 — Method

The path of the ATmega32U4 file definition.

AtmelToolsDeviceFiles.AtmelLibrary.ATmega406 — Method

The path of the ATmega406 file definition.

AtmelToolsDeviceFiles.AtmelLibrary.ATmega48 — Method

The path of the ATmega48 file definition.

AtmelToolsDeviceFiles.AtmelLibrary.ATmega4808 — Method

The path of the ATmega4808 file definition.

AtmelToolsDeviceFiles.AtmelLibrary.ATmega4809 — Method

The path of the ATmega4809 file definition.

AtmelToolsDeviceFiles.AtmelLibrary.ATmega48A — Method

The path of the ATmega48A file definition.

AtmelToolsDeviceFiles.AtmelLibrary.ATmega48P — Method

The path of the ATmega48P file definition.

AtmelToolsDeviceFiles.AtmelLibrary.ATmega48PA — Method

The path of the ATmega48PA file definition.

AtmelToolsDeviceFiles.AtmelLibrary.ATmega48PB — Method

The path of the ATmega48PB file definition.

AtmelToolsDeviceFiles.AtmelLibrary.ATmega64 — Method

The path of the ATmega64 file definition.

AtmelToolsDeviceFiles.AtmelLibrary.ATmega640 — Method

The path of the ATmega640 file definition.

AtmelToolsDeviceFiles.AtmelLibrary.ATmega644 — Method

The path of the ATmega644 file definition.

AtmelToolsDeviceFiles.AtmelLibrary.ATmega644A — Method

The path of the ATmega644A file definition.

AtmelToolsDeviceFiles.AtmelLibrary.ATmega644P — Method

The path of the ATmega644P file definition.

AtmelToolsDeviceFiles.AtmelLibrary.ATmega644PA — Method

The path of the ATmega644PA file definition.

AtmelToolsDeviceFiles.AtmelLibrary.ATmega644RFR2 — Method

The path of the ATmega644RFR2 file definition.

AtmelToolsDeviceFiles.AtmelLibrary.ATmega645 — Method

The path of the ATmega645 file definition.

AtmelToolsDeviceFiles.AtmelLibrary.ATmega6450 — Method

The path of the ATmega6450 file definition.

AtmelToolsDeviceFiles.AtmelLibrary.ATmega6450A — Method

The path of the ATmega6450A file definition.

AtmelToolsDeviceFiles.AtmelLibrary.ATmega6450P — Method

The path of the ATmega6450P file definition.

AtmelToolsDeviceFiles.AtmelLibrary.ATmega645A — Method

The path of the ATmega645A file definition.

AtmelToolsDeviceFiles.AtmelLibrary.ATmega645P — Method

The path of the ATmega645P file definition.

AtmelToolsDeviceFiles.AtmelLibrary.ATmega649 — Method

The path of the ATmega649 file definition.

AtmelToolsDeviceFiles.AtmelLibrary.ATmega6490 — Method

The path of the ATmega6490 file definition.

AtmelToolsDeviceFiles.AtmelLibrary.ATmega6490A — Method

The path of the ATmega6490A file definition.

AtmelToolsDeviceFiles.AtmelLibrary.ATmega6490P — Method

The path of the ATmega6490P file definition.

AtmelToolsDeviceFiles.AtmelLibrary.ATmega649A — Method

The path of the ATmega649A file definition.

AtmelToolsDeviceFiles.AtmelLibrary.ATmega649P — Method

The path of the ATmega649P file definition.

AtmelToolsDeviceFiles.AtmelLibrary.ATmega64A — Method

The path of the ATmega64A file definition.

AtmelToolsDeviceFiles.AtmelLibrary.ATmega64C1 — Method

The path of the ATmega64C1 file definition.

AtmelToolsDeviceFiles.AtmelLibrary.ATmega64HVE2 — Method

The path of the ATmega64HVE2 file definition.

AtmelToolsDeviceFiles.AtmelLibrary.ATmega64M1 — Method

The path of the ATmega64M1 file definition.

AtmelToolsDeviceFiles.AtmelLibrary.ATmega64RFR2 — Method

The path of the ATmega64RFR2 file definition.

AtmelToolsDeviceFiles.AtmelLibrary.ATmega8 — Method

The path of the ATmega8 file definition.

AtmelToolsDeviceFiles.AtmelLibrary.ATmega808 — Method

The path of the ATmega808 file definition.

AtmelToolsDeviceFiles.AtmelLibrary.ATmega809 — Method

The path of the ATmega809 file definition.

AtmelToolsDeviceFiles.AtmelLibrary.ATmega8515 — Method

The path of the ATmega8515 file definition.

AtmelToolsDeviceFiles.AtmelLibrary.ATmega8535 — Method

The path of the ATmega8535 file definition.

AtmelToolsDeviceFiles.AtmelLibrary.ATmega88 — Method

The path of the ATmega88 file definition.

AtmelToolsDeviceFiles.AtmelLibrary.ATmega88A — Method

The path of the ATmega88A file definition.

AtmelToolsDeviceFiles.AtmelLibrary.ATmega88P — Method

The path of the ATmega88P file definition.

AtmelToolsDeviceFiles.AtmelLibrary.ATmega88PA — Method

The path of the ATmega88PA file definition.

AtmelToolsDeviceFiles.AtmelLibrary.ATmega88PB — Method

The path of the ATmega88PB file definition.

AtmelToolsDeviceFiles.AtmelLibrary.ATmega8A — Method

The path of the ATmega8A file definition.

AtmelToolsDeviceFiles.AtmelLibrary.ATmega8HVA — Method

The path of the ATmega8HVA file definition.

AtmelToolsDeviceFiles.AtmelLibrary.ATmega8U2 — Method

The path of the ATmega8U2 file definition.

AtmelToolsDeviceFiles.AtmelLibrary.ATtiny10 — Method

The path of the ATtiny10 file definition.

AtmelToolsDeviceFiles.AtmelLibrary.ATtiny102 — Method

The path of the ATtiny102 file definition.

AtmelToolsDeviceFiles.AtmelLibrary.ATtiny104 — Method

The path of the ATtiny104 file definition.

AtmelToolsDeviceFiles.AtmelLibrary.ATtiny11 — Method

The path of the ATtiny11 file definition.

AtmelToolsDeviceFiles.AtmelLibrary.ATtiny12 — Method

The path of the ATtiny12 file definition.

AtmelToolsDeviceFiles.AtmelLibrary.ATtiny13 — Method

The path of the ATtiny13 file definition.

AtmelToolsDeviceFiles.AtmelLibrary.ATtiny13A — Method

The path of the ATtiny13A file definition.

AtmelToolsDeviceFiles.AtmelLibrary.ATtiny15 — Method

The path of the ATtiny15 file definition.

AtmelToolsDeviceFiles.AtmelLibrary.ATtiny1604 — Method

The path of the ATtiny1604 file definition.

AtmelToolsDeviceFiles.AtmelLibrary.ATtiny1606 — Method

The path of the ATtiny1606 file definition.

AtmelToolsDeviceFiles.AtmelLibrary.ATtiny1607 — Method

The path of the ATtiny1607 file definition.

AtmelToolsDeviceFiles.AtmelLibrary.ATtiny1614 — Method

The path of the ATtiny1614 file definition.

AtmelToolsDeviceFiles.AtmelLibrary.ATtiny1616 — Method

The path of the ATtiny1616 file definition.

AtmelToolsDeviceFiles.AtmelLibrary.ATtiny1617 — Method

The path of the ATtiny1617 file definition.

AtmelToolsDeviceFiles.AtmelLibrary.ATtiny1624 — Method

The path of the ATtiny1624 file definition.

AtmelToolsDeviceFiles.AtmelLibrary.ATtiny1626 — Method

The path of the ATtiny1626 file definition.

AtmelToolsDeviceFiles.AtmelLibrary.ATtiny1627 — Method

The path of the ATtiny1627 file definition.

AtmelToolsDeviceFiles.AtmelLibrary.ATtiny1634 — Method

The path of the ATtiny1634 file definition.

AtmelToolsDeviceFiles.AtmelLibrary.ATtiny167 — Method

The path of the ATtiny167 file definition.

AtmelToolsDeviceFiles.AtmelLibrary.ATtiny20 — Method

The path of the ATtiny20 file definition.

AtmelToolsDeviceFiles.AtmelLibrary.ATtiny202 — Method

The path of the ATtiny202 file definition.

AtmelToolsDeviceFiles.AtmelLibrary.ATtiny204 — Method

The path of the ATtiny204 file definition.

AtmelToolsDeviceFiles.AtmelLibrary.ATtiny212 — Method

The path of the ATtiny212 file definition.

AtmelToolsDeviceFiles.AtmelLibrary.ATtiny214 — Method

The path of the ATtiny214 file definition.

AtmelToolsDeviceFiles.AtmelLibrary.ATtiny2313 — Method

The path of the ATtiny2313 file definition.

AtmelToolsDeviceFiles.AtmelLibrary.ATtiny2313A — Method

The path of the ATtiny2313A file definition.

AtmelToolsDeviceFiles.AtmelLibrary.ATtiny24 — Method

The path of the ATtiny24 file definition.

AtmelToolsDeviceFiles.AtmelLibrary.ATtiny24A — Method

The path of the ATtiny24A file definition.

AtmelToolsDeviceFiles.AtmelLibrary.ATtiny25 — Method

The path of the ATtiny25 file definition.

AtmelToolsDeviceFiles.AtmelLibrary.ATtiny26 — Method

The path of the ATtiny26 file definition.

AtmelToolsDeviceFiles.AtmelLibrary.ATtiny261 — Method

The path of the ATtiny261 file definition.

AtmelToolsDeviceFiles.AtmelLibrary.ATtiny261A — Method

The path of the ATtiny261A file definition.

AtmelToolsDeviceFiles.AtmelLibrary.ATtiny3216 — Method

The path of the ATtiny3216 file definition.

AtmelToolsDeviceFiles.AtmelLibrary.ATtiny3217 — Method

The path of the ATtiny3217 file definition.

AtmelToolsDeviceFiles.AtmelLibrary.ATtiny3224 — Method

The path of the ATtiny3224 file definition.

AtmelToolsDeviceFiles.AtmelLibrary.ATtiny3226 — Method

The path of the ATtiny3226 file definition.

AtmelToolsDeviceFiles.AtmelLibrary.ATtiny3227 — Method

The path of the ATtiny3227 file definition.

AtmelToolsDeviceFiles.AtmelLibrary.ATtiny4 — Method

The path of the ATtiny4 file definition.

AtmelToolsDeviceFiles.AtmelLibrary.ATtiny40 — Method

The path of the ATtiny40 file definition.

AtmelToolsDeviceFiles.AtmelLibrary.ATtiny402 — Method

The path of the ATtiny402 file definition.

AtmelToolsDeviceFiles.AtmelLibrary.ATtiny404 — Method

The path of the ATtiny404 file definition.

AtmelToolsDeviceFiles.AtmelLibrary.ATtiny406 — Method

The path of the ATtiny406 file definition.

AtmelToolsDeviceFiles.AtmelLibrary.ATtiny412 — Method

The path of the ATtiny412 file definition.

AtmelToolsDeviceFiles.AtmelLibrary.ATtiny414 — Method

The path of the ATtiny414 file definition.

AtmelToolsDeviceFiles.AtmelLibrary.ATtiny416 — Method

The path of the ATtiny416 file definition.

AtmelToolsDeviceFiles.AtmelLibrary.ATtiny417 — Method

The path of the ATtiny417 file definition.

AtmelToolsDeviceFiles.AtmelLibrary.ATtiny424 — Method

The path of the ATtiny424 file definition.

AtmelToolsDeviceFiles.AtmelLibrary.ATtiny426 — Method

The path of the ATtiny426 file definition.

AtmelToolsDeviceFiles.AtmelLibrary.ATtiny427 — Method

The path of the ATtiny427 file definition.

AtmelToolsDeviceFiles.AtmelLibrary.ATtiny4313 — Method

The path of the ATtiny4313 file definition.

AtmelToolsDeviceFiles.AtmelLibrary.ATtiny43U — Method

The path of the ATtiny43U file definition.

AtmelToolsDeviceFiles.AtmelLibrary.ATtiny44 — Method

The path of the ATtiny44 file definition.

AtmelToolsDeviceFiles.AtmelLibrary.ATtiny441 — Method

The path of the ATtiny441 file definition.

AtmelToolsDeviceFiles.AtmelLibrary.ATtiny44A — Method

The path of the ATtiny44A file definition.

AtmelToolsDeviceFiles.AtmelLibrary.ATtiny45 — Method

The path of the ATtiny45 file definition.

AtmelToolsDeviceFiles.AtmelLibrary.ATtiny461 — Method

The path of the ATtiny461 file definition.

AtmelToolsDeviceFiles.AtmelLibrary.ATtiny461A — Method

The path of the ATtiny461A file definition.

AtmelToolsDeviceFiles.AtmelLibrary.ATtiny48 — Method

The path of the ATtiny48 file definition.

AtmelToolsDeviceFiles.AtmelLibrary.ATtiny5 — Method

The path of the ATtiny5 file definition.

AtmelToolsDeviceFiles.AtmelLibrary.ATtiny804 — Method

The path of the ATtiny804 file definition.

AtmelToolsDeviceFiles.AtmelLibrary.ATtiny806 — Method

The path of the ATtiny806 file definition.

AtmelToolsDeviceFiles.AtmelLibrary.ATtiny807 — Method

The path of the ATtiny807 file definition.

AtmelToolsDeviceFiles.AtmelLibrary.ATtiny814 — Method

The path of the ATtiny814 file definition.

AtmelToolsDeviceFiles.AtmelLibrary.ATtiny816 — Method

The path of the ATtiny816 file definition.

AtmelToolsDeviceFiles.AtmelLibrary.ATtiny817 — Method

The path of the ATtiny817 file definition.

AtmelToolsDeviceFiles.AtmelLibrary.ATtiny824 — Method

The path of the ATtiny824 file definition.

AtmelToolsDeviceFiles.AtmelLibrary.ATtiny826 — Method

The path of the ATtiny826 file definition.

AtmelToolsDeviceFiles.AtmelLibrary.ATtiny827 — Method

The path of the ATtiny827 file definition.

AtmelToolsDeviceFiles.AtmelLibrary.ATtiny828 — Method

The path of the ATtiny828 file definition.

AtmelToolsDeviceFiles.AtmelLibrary.ATtiny84 — Method

The path of the ATtiny84 file definition.

AtmelToolsDeviceFiles.AtmelLibrary.ATtiny841 — Method

The path of the ATtiny841 file definition.

AtmelToolsDeviceFiles.AtmelLibrary.ATtiny84A — Method

The path of the ATtiny84A file definition.

AtmelToolsDeviceFiles.AtmelLibrary.ATtiny85 — Method

The path of the ATtiny85 file definition.

AtmelToolsDeviceFiles.AtmelLibrary.ATtiny861 — Method

The path of the ATtiny861 file definition.

AtmelToolsDeviceFiles.AtmelLibrary.ATtiny861A — Method

The path of the ATtiny861A file definition.

AtmelToolsDeviceFiles.AtmelLibrary.ATtiny87 — Method

The path of the ATtiny87 file definition.

AtmelToolsDeviceFiles.AtmelLibrary.ATtiny88 — Method

The path of the ATtiny88 file definition.

AtmelToolsDeviceFiles.AtmelLibrary.ATtiny9 — Method

The path of the ATtiny9 file definition.

Internals

AtmelToolsDeviceFiles.AtmelToolsDeviceFilePath — Type

A convenience wrapper for the path to a device definition file.

AtmelToolsDeviceFiles.devicefilepath — Function

Returns the path to the devicde definition file.