Atmel Tools Device Files

Atmel Tools Device Files (.atdf) are parsed to produce a specific set of typed objects listed here.

Atmel Tools Device Files

Device variants

Device summary

Peripheral modules

Pinouts

Index

AtmelToolsDeviceFiles.AVRToolsDeviceFile — Type

A fully parsed Atmel Tools Device File.

variants::Vector{AtmelToolsDeviceFiles.Variant}: See also Variant.

devices::Vector{AtmelToolsDeviceFiles.Device}: See also Device.

modules::Vector{AtmelToolsDeviceFiles.PeripheralModule}: See also PeripheralModule.

pinouts::Vector{AtmelToolsDeviceFiles.Pinout}: See also Pinout.

Device variants

AtmelToolsDeviceFiles.Variant — Type

Describes a specific package for a device.

ordercode::String

package::String

pinout::Union{Nothing, String}

speedmax::Int64

tempmax::Float64

tempmin::Float64

vccmax::Float64

vccmin::Float64

Device summary

This section summarizes a device's architecture, family, name, and memory layout.

AtmelToolsDeviceFiles.Device — Type

Describes a device

architecture::String

family::String

name::String

address_spaces::Vector{AtmelToolsDeviceFiles.AdressSpace}: See also AdressSpace.

peripherals::Vector{AtmelToolsDeviceFiles.DeviceModule}: See also DeviceModule.

interrupts::Vector{AtmelToolsDeviceFiles.Interrupt}: See also Interrupt.

interfaces::Vector{AtmelToolsDeviceFiles.Interface}: See also Interface.

property_groups::Vector{AtmelToolsDeviceFiles.PropertyGroup}: See also PropertyGroup.

parameters::Vector{AtmelToolsDeviceFiles.Parameter}: See also Parameter.

AtmelToolsDeviceFiles.Parameter — Type

Describes a parameter.

name::String

value::String

AtmelToolsDeviceFiles.AdressSpace — Type

Describes an adress space

endianness::String

id::String

name::String

size::UInt64

start::UInt64

memory_segments::Vector{AtmelToolsDeviceFiles.MemorySegment}: See also MemorySegment.

AtmelToolsDeviceFiles.MemorySegment — Type

Describes a memory segment.

exec::Union{Nothing, String}

name::String

rw::Union{Nothing, String}

size::UInt64

start::UInt64

type::String

AtmelToolsDeviceFiles.DeviceModule — Type

Describes a peripheral module.

id::Union{Nothing, String}

name::String

instances::Vector{AtmelToolsDeviceFiles.Instance}: See also Instance.

AtmelToolsDeviceFiles.Instance — Type

Describes an instance of a peripheral module.

name::String

register_group::Union{Nothing, AtmelToolsDeviceFiles.RegisterGroup}: See also RegisterGroup.

signals::Vector{AtmelToolsDeviceFiles.Signal}: See also Signal.

parameters::Vector{AtmelToolsDeviceFiles.Parameter}: See also Parameter.

AtmelToolsDeviceFiles.RegisterGroup — Type

Describes the register group used by a peripheral instance.

address_space::String

name::String

name_in_module::String

offset::UInt64

AtmelToolsDeviceFiles.Signal — Type

Describes a signal for a peripheral module.

function_::Union{Nothing, String}

group::String

index::Union{Nothing, UInt64}

pad::String

field::Union{Nothing, String}

AtmelToolsDeviceFiles.Interrupt — Type

Describes an interrupt.

index::Int64

module_instance::Union{Nothing, String}

name::String

AtmelToolsDeviceFiles.Interface — Type

Describes an interface to a device.

name::String

type::String

AtmelToolsDeviceFiles.PropertyGroup — Type

Describes a property group for a device.

name::String

properties::Vector{AtmelToolsDeviceFiles.Property}

AtmelToolsDeviceFiles.Property — Type

Describes a property.

name::String

value::String

Peripheral modules

This contains the extended description of available peripheral modules.

AtmelToolsDeviceFiles.PeripheralModule — Type

Describes the layout and usage of a peripheral module.

caption::Union{Nothing, String}

id::Union{Nothing, String}

name::String

register_group::Vector{AtmelToolsDeviceFiles.PeripheralRegisterGroup}: See also PeripheralRegisterGroup.

value_groups::Vector{AtmelToolsDeviceFiles.PeripheralValueGroup}: See also PeripheralValueGroup.

The modules listed here correspond to the short version listed as DeviceModule in the device definition.

AtmelToolsDeviceFiles.PeripheralRegisterGroup — Type

Describes a group of registers related to a peripheral.

caption::Union{Nothing, String}

name::String

size::Union{Nothing, UInt64}

registers::Vector{AtmelToolsDeviceFiles.PeripheralRegister}: See also PeripheralRegister.

modes::Vector{AtmelToolsDeviceFiles.PeripheralRegisterGroupMode}: See also PeripheralRegisterGroupMode.

AtmelToolsDeviceFiles.PeripheralRegisterGroupMode — Type

Mode of a register group

caption::String

name::String

qualifier::String

value::String

AtmelToolsDeviceFiles.PeripheralRegister — Type

Describes a register related to a peripheral.

caption::Union{Nothing, String}

initval::Union{Nothing, UInt64}

name::String

offset::UInt64

rw::Union{Nothing, String}

size::Int64

bitfields::Vector{AtmelToolsDeviceFiles.PeripheralBitField}: See also PeripheralBitField.

modes::Vector{AtmelToolsDeviceFiles.RegisterMode}: See also RegisterMode.

AtmelToolsDeviceFiles.RegisterMode — Type

Register modes.

name::String

qualifier::String

value::String

AtmelToolsDeviceFiles.PeripheralBitField — Type

Describes a bit field in a register associated to a peripheral.

caption::Union{Nothing, String}

mask::UInt64

name::String

rw::Union{Nothing, String}

values::Union{Nothing, String}

AtmelToolsDeviceFiles.PeripheralValueGroup — Type

Describes a set of values for a peripheral module.

caption::Union{Nothing, String}

name::String

values::Vector{AtmelToolsDeviceFiles.PeripheralValue}: See also PeripheralValue

AtmelToolsDeviceFiles.PeripheralValue — Type

Describes a specific value for a peripheral module.

caption::String

name::String

value::String

Pinouts

This describes the available pinouts for a device.

AtmelToolsDeviceFiles.Pinout — Type

Describe a specific available pinout for a device.

name::String

pins::Vector{AtmelToolsDeviceFiles.Pin}: See also Pin.

AtmelToolsDeviceFiles.Pin — Type

Describes a pin for a device's pinout.

pad::String

position::String

Index

AtmelToolsDeviceFiles.AVRToolsDeviceFile

AtmelToolsDeviceFiles.AdressSpace

AtmelToolsDeviceFiles.Device

AtmelToolsDeviceFiles.DeviceModule

AtmelToolsDeviceFiles.Instance

AtmelToolsDeviceFiles.Interface

AtmelToolsDeviceFiles.Interrupt

AtmelToolsDeviceFiles.MemorySegment

AtmelToolsDeviceFiles.Parameter

AtmelToolsDeviceFiles.PeripheralBitField

AtmelToolsDeviceFiles.PeripheralModule

AtmelToolsDeviceFiles.PeripheralRegister

AtmelToolsDeviceFiles.PeripheralRegisterGroup

AtmelToolsDeviceFiles.PeripheralRegisterGroupMode

AtmelToolsDeviceFiles.PeripheralValue

AtmelToolsDeviceFiles.PeripheralValueGroup

AtmelToolsDeviceFiles.Pin

AtmelToolsDeviceFiles.Pinout

AtmelToolsDeviceFiles.Property

AtmelToolsDeviceFiles.PropertyGroup

AtmelToolsDeviceFiles.RegisterGroup

AtmelToolsDeviceFiles.RegisterMode

AtmelToolsDeviceFiles.Signal

AtmelToolsDeviceFiles.Variant